+++ source/modulo_utils/include/modulo_utils/parsing.hpp
@@ -4,7 +4,23 @@
 
 #include <rclcpp/rclcpp.hpp>
 
-namespace modulo_utils {
+/**
+ * @namespace modulo_utils::parsing
+ * @brief Modulo parsing helpers.
+ */
+namespace modulo_utils::parsing {
+
+/**
+ * @brief This function returns a default validation warning for topic names.
+ * @param name The pre-validation signal name
+ * @param type One of input|output
+ * @return The validation warning
+ * @see modulo_utils::parsing::parse_topic_name
+*/
+[[maybe_unused]] static const std::string topic_validation_warning(const std::string& name, const std::string& type) {
+  return "The parsed signal name for " + type + " '" + name
+      + "' is empty. Provide a string with valid characters for the signal name ([a-z0-9_]).";
+}
 
 /**
  * @brief Parse a string node name from NodeOptions arguments.
@@ -29,17 +45,16 @@ parse_node_name(const rclcpp::NodeOptions& options, const std::string& fallback 
 
 /**
  * @brief Parse a string topic name from a user-provided input.
- * @details This functions removes all characters different from a-z, A-Z, 0-9, and _ from a string and transforms
- * uppercase letters into lowercase letters. Additionally, it removes all leading numbers and underscores, such that the
- * resulting string starts with a letter a-z.
+ * @details This functions removes all characters different from a-z, 0-9, and _ from a string. Additionally, it removes
+ * all leading numbers and underscores, such that the resulting string starts with a letter a-z.
  * @param topic_name The input string
  * @return the sanitized string
  */
 [[maybe_unused]] static std::string parse_topic_name(const std::string& topic_name) {
   std::string output;
   for (char c : topic_name) {
-    if ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z')) {
-      output.insert(output.end(), std::tolower(c));
+    if (c >= 'a' && c <= 'z') {
+      output.insert(output.end(), c);
     } else if (!output.empty() && ((c >= '0' && c <= '9') || c == '_')) {
       output.insert(output.end(), c);
     }
@@ -47,4 +62,4 @@ parse_node_name(const rclcpp::NodeOptions& options, const std::string& fallback 
   return output;
 }
 
-}// namespace modulo_utils
+}// namespace modulo_utils::parsing